announcements-server-py/requirements.txt
requests==2.32.3
pandas==2.2.3
schedule==1.2.2
python-dotenv==1.0.1
PyMuPDF==1.24.13
qdrant-client==1.12.1
fastembed==0.4.2
pymongo==4.10.1
